Always use commit sha1 for QML_COMPILE_HASH
git describe can produce different strings for the same sha1, depending
on the version of git (length of short hash) and whether release tag was
in place when build was done (v5.11.1-0- vs. v5.11.0-66-).

Instead use rev-parse to get the unambiguous commit sha1 which is also
what is used from the .tag file.

Change-Id: If378f03eb764e2d858bda5b662106aa4402d6559
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>